requirements.txt
torch==2.2.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
spotipy==2.23.0
python-dotenv==1.0.1